DragDiffusion_edited/condaenv.e61xikby.requirements.txt
torch==2.0.0
torchvision==0.15.1
gradio==3.41.1
pydantic==2.0.2
albumentations==1.3.0
opencv-contrib-python==4.3.0.36
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.5.0
omegaconf==2.3.0
test-tube>=0.7.5
streamlit==1.12.1
einops==0.6.0
transformers==4.27.0
webdataset==0.2.5
kornia==0.6
open_clip_torch==2.16.0
invisible-watermark>=0.1.5
streamlit-drawable-canvas==0.8.0
torchmetrics==0.6.0
timm==0.6.12
addict==2.4.0
yapf==0.32.0
prettytable==3.6.0
safetensors==0.3.1
basicsr==1.4.2
accelerate==0.17.0
decord==0.6.0
diffusers==0.24.0
moviepy==1.0.3
opencv_python==4.7.0.68
Pillow==9.4.0
scikit_image==0.19.3
scipy==1.10.1
tensorboardX==2.6
tqdm==4.64.1
numpy==1.24.1